+++ common/src/main/java/io/github/rulft44/druids/client/DruidsClient.java
@@ -3,25 +3,30 @@ package io.github.rulft44.druids.client;
 import io.github.rulft44.druids.Druids;
 import io.github.rulft44.druids.client.armor.DruidArmorRenderer;
 import io.github.rulft44.druids.item.ModArmors;
+import io.github.rulft44.druids.spell.skill.DruidsSkillDefinitions;
+import io.github.rulft44.druids.utils.DruidsTranslationUtils;
 import mod.azure.azurelibarmor.common.render.armor.AzArmorRenderer;
 import mod.azure.azurelibarmor.common.render.armor.AzArmorRendererRegistry;
 import net.minecraft.util.Identifier;
-import net.spell_engine.api.item.armor.Armor;
-import net.spell_engine.api.render.CustomModels;
+import net.skill_tree_rpgs.utils.TranslationUtil;
+import net.spell_engine.rpg_series.item.Armor;
 
-import java.util.List;
 import java.util.function.Supplier;
 
 public class DruidsClient{
 	public static void init() {
-		CustomModels.registerModelIds(List.of(
-			Identifier.of(Druids.ID, "projectile/bramble_shot"),
-			Identifier.of(Druids.ID, "effect/entanglement1"),
-			Identifier.of(Druids.ID, "effect/entanglement11"),
-			Identifier.of(Druids.ID, "effect/entanglement2"),
-			Identifier.of(Druids.ID, "effect/dandelion"),
-			Identifier.of(Druids.ID, "effect/poppy")
-		));
+		for (var entry: DruidsSkillDefinitions.ENTRIES) {
+			var skillId = entry.id();
+			if (entry.spellReward() != null) {
+				var container = entry.spellReward().get(0);
+				var id = Identifier.of(container.spell_ids().getFirst());
+				TranslationUtil.resolvers.put(skillId, () -> TranslationUtil.resolveSpellDetails(id));
+			}
+			else if (entry.attributeReward() != null) {
+				var attribute = entry.attributeReward();
+				TranslationUtil.resolvers.put(skillId, () -> DruidsTranslationUtils.resolveAttributeModifierTooltip(attribute));
+			}
+		}
 
 		registerArmorRenderer(ModArmors.druidArmorSet, DruidArmorRenderer::druid);
 		registerArmorRenderer(ModArmors.netheriteDruidArmorSet, DruidArmorRenderer::netherite_druid);