+++ common/src/main/java/io/github/rulft44/druids/config/TweaksConfig.java
@@ -2,4 +2,5 @@ package io.github.rulft44.druids.config;
 
 public class TweaksConfig {
 	public boolean ignore_items_required_mods = false;
+	public boolean disable_druids_skilltree_changes = false;
 }