+++ common/src/main/java/io/github/rulft44/druids/item/armor/DruidArmor.java
@@ -2,7 +2,7 @@ package io.github.rulft44.druids.item.armor;
 
 import net.minecraft.item.ArmorMaterial;
 import net.minecraft.registry.entry.RegistryEntry;
-import net.spell_engine.api.item.armor.Armor;
+import net.spell_engine.rpg_series.item.Armor;
 
 public class DruidArmor extends Armor.CustomItem{
 	public DruidArmor(RegistryEntry<ArmorMaterial> material, Type slot, Settings settings) {