+++ fabric/src/main/java/io/github/rulft44/fabric/FabricMod.java
@@ -2,6 +2,12 @@ package io.github.rulft44.fabric;
 
 import io.github.rulft44.druids.Druids;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
+import net.fabricmc.fabric.api.resource.ResourcePackActivationType;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.util.Identifier;
+
+import static io.github.rulft44.druids.Druids.ID;
 
 public final class FabricMod implements ModInitializer {
     @Override
@@ -10,8 +16,17 @@ public final class FabricMod implements ModInitializer {
         // However, some things (like resources) may still be uninitialized.
         // Proceed with mild caution.
 
+		FabricLoader.getInstance().getModContainer(ID).ifPresent(modContainer -> {
+			ResourceManagerHelper.registerBuiltinResourcePack(
+				Identifier.of(ID, "druids_skill_tree_changes"),
+				modContainer,
+				ResourcePackActivationType.ALWAYS_ENABLED
+			);
+		});
+
         // Run our common setup.
         Druids.init();
+		Druids.registerResourcePack();
 		Druids.registerItems();
 		Druids.registerSounds();
 		Druids.registerEffects();