+++ fabric/src/main/java/io/github/rulft44/fabric/datagen/DruidRecipes.java
@@ -12,7 +12,7 @@ import net.minecraft.item.Items;
 import net.minecraft.recipe.book.RecipeCategory;
 import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.registry.tag.ItemTags;
-import net.spell_engine.api.item.armor.Armor;
+import net.spell_engine.rpg_series.item.Armor;
 
 import java.util.concurrent.CompletableFuture;